# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Oficina Mecânica Management System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cliente {
  id: number;
  nome: string;
  telefone: string;
  whatsapp?: string;
  email?: string;
  observacoes?: string;
  createdAt: string;
}

export interface CreateCliente {
  nome: string;
  telefone: string;
  whatsapp?: string;
  email?: string;
  observacoes?: string;
}

export interface Veiculo {
  id: number;
  clienteId: number;
  clienteNome?: string;
  placa: string;
  modelo: string;
  marca: string;
  ano: number;
  km?: number;
  cor?: string;
  observacoes?: string;
  createdAt: string;
}

export interface CreateVeiculo {
  clienteId: number;
  placa: string;
  modelo: string;
  marca: string;
  ano: number;
  km?: number;
  cor?: string;
  observacoes?: string;
}

export interface Peca {
  id: number;
  nome: string;
  codigo?: string;
  quantidade: number;
  quantidadeMinima: number;
  valorCusto: number;
  valorVenda: number;
  estoqueAlerta: boolean;
  createdAt: string;
}

export interface CreatePeca {
  nome: string;
  codigo?: string;
  quantidade: number;
  quantidadeMinima: number;
  valorCusto: number;
  valorVenda: number;
}

export type MovimentacaoPecaTipo =
  (typeof MovimentacaoPecaTipo)[keyof typeof MovimentacaoPecaTipo];

export const MovimentacaoPecaTipo = {
  entrada: "entrada",
  saida: "saida",
} as const;

export interface MovimentacaoPeca {
  tipo: MovimentacaoPecaTipo;
  quantidade: number;
  motivo?: string;
}

export interface Servico {
  id: number;
  nome: string;
  descricao?: string;
  valorPadrao: number;
  createdAt: string;
}

export interface CreateServico {
  nome: string;
  descricao?: string;
  valorPadrao: number;
}

export type OrdemServicoStatus =
  (typeof OrdemServicoStatus)[keyof typeof OrdemServicoStatus];

export const OrdemServicoStatus = {
  orcamento: "orcamento",
  em_andamento: "em_andamento",
  finalizado: "finalizado",
  entregue: "entregue",
} as const;

export interface OrdemServico {
  id: number;
  numero: string;
  clienteId: number;
  clienteNome: string;
  veiculoId: number;
  veiculoPlaca: string;
  veiculoModelo: string;
  responsavel: string;
  status: OrdemServicoStatus;
  dataEntrada: string;
  dataPrevisao?: string;
  dataFinalizacao?: string;
  observacoes?: string;
  totalServicos: number;
  totalPecas: number;
  total: number;
  atrasado: boolean;
  createdAt: string;
}

export type OrdemServicoDetalhadaStatus =
  (typeof OrdemServicoDetalhadaStatus)[keyof typeof OrdemServicoDetalhadaStatus];

export const OrdemServicoDetalhadaStatus = {
  orcamento: "orcamento",
  em_andamento: "em_andamento",
  finalizado: "finalizado",
  entregue: "entregue",
} as const;

export type OrdemServicoDetalhadaChecklistEntrada = { [key: string]: unknown };

export type OrdemServicoDetalhadaChecklistEntrega = { [key: string]: unknown };

export interface OrdemServicoItem {
  id: number;
  servicoId: number;
  nome: string;
  valor: number;
}

export interface OrdemPecaItem {
  id: number;
  pecaId: number;
  nome: string;
  codigo?: string;
  quantidade: number;
  valorUnitario: number;
  valorTotal: number;
}

export interface OrdemServicoDetalhada {
  id: number;
  numero: string;
  clienteId: number;
  clienteNome: string;
  veiculoId: number;
  veiculoPlaca: string;
  veiculoModelo: string;
  veiculoMarca: string;
  veiculoAno: number;
  veiculoKm?: number;
  responsavel: string;
  status: OrdemServicoDetalhadaStatus;
  dataEntrada: string;
  dataPrevisao?: string;
  dataFinalizacao?: string;
  observacoes?: string;
  checklistEntrada?: OrdemServicoDetalhadaChecklistEntrada;
  checklistEntrega?: OrdemServicoDetalhadaChecklistEntrega;
  servicos: OrdemServicoItem[];
  pecas: OrdemPecaItem[];
  totalServicos: number;
  totalPecas: number;
  total: number;
  atrasado: boolean;
  createdAt: string;
}

export type CreateOrdemServicoStatus =
  (typeof CreateOrdemServicoStatus)[keyof typeof CreateOrdemServicoStatus];

export const CreateOrdemServicoStatus = {
  orcamento: "orcamento",
  em_andamento: "em_andamento",
  finalizado: "finalizado",
  entregue: "entregue",
} as const;

export type CreateOrdemServicoChecklistEntrada = { [key: string]: unknown };

export interface AddServicoItem {
  servicoId: number;
  valor: number;
}

export interface AddPecaItem {
  pecaId: number;
  quantidade: number;
  valorUnitario: number;
}

export interface CreateOrdemServico {
  clienteId: number;
  veiculoId: number;
  responsavel: string;
  status?: CreateOrdemServicoStatus;
  dataEntrada: string;
  dataPrevisao?: string;
  observacoes?: string;
  checklistEntrada?: CreateOrdemServicoChecklistEntrada;
  servicos?: AddServicoItem[];
  pecas?: AddPecaItem[];
}

export type UpdateOrdemServicoStatus =
  (typeof UpdateOrdemServicoStatus)[keyof typeof UpdateOrdemServicoStatus];

export const UpdateOrdemServicoStatus = {
  orcamento: "orcamento",
  em_andamento: "em_andamento",
  finalizado: "finalizado",
  entregue: "entregue",
} as const;

export type UpdateOrdemServicoChecklistEntrada = { [key: string]: unknown };

export type UpdateOrdemServicoChecklistEntrega = { [key: string]: unknown };

export interface UpdateOrdemServico {
  clienteId?: number;
  veiculoId?: number;
  responsavel?: string;
  status?: UpdateOrdemServicoStatus;
  dataEntrada?: string;
  dataPrevisao?: string;
  dataFinalizacao?: string;
  observacoes?: string;
  checklistEntrada?: UpdateOrdemServicoChecklistEntrada;
  checklistEntrega?: UpdateOrdemServicoChecklistEntrega;
  servicos?: AddServicoItem[];
  pecas?: AddPecaItem[];
}

export type UpdateStatusStatus =
  (typeof UpdateStatusStatus)[keyof typeof UpdateStatusStatus];

export const UpdateStatusStatus = {
  orcamento: "orcamento",
  em_andamento: "em_andamento",
  finalizado: "finalizado",
  entregue: "entregue",
} as const;

export type UpdateStatusChecklistEntrega = { [key: string]: unknown };

export interface UpdateStatus {
  status: UpdateStatusStatus;
  checklistEntrega?: UpdateStatusChecklistEntrega;
}

export interface Despesa {
  id: number;
  descricao: string;
  valor: number;
  categoria: string;
  data: string;
  createdAt: string;
}

export interface CreateDespesa {
  descricao: string;
  valor: number;
  categoria: string;
  data: string;
}

export interface ResumoFinanceiro {
  totalReceitas: number;
  totalDespesas: number;
  lucro: number;
  ordensFinalizadas: number;
  ticketMedio: number;
}

export interface FluxoCaixaDia {
  data: string;
  receitas: number;
  despesas: number;
  saldo: number;
}

export interface Dashboard {
  ordensEmAndamento: number;
  ordensAtrasadas: number;
  ordensHoje: number;
  faturamentoDia: number;
  faturamentoMes: number;
  pecasEstoqueBaixo: number;
  ordens: OrdemServico[];
}

export type ListVeiculosParams = {
  clienteId?: number;
};

export type ListOrdensParams = {
  status?: ListOrdensStatus;
};

export type ListOrdensStatus =
  (typeof ListOrdensStatus)[keyof typeof ListOrdensStatus];

export const ListOrdensStatus = {
  orcamento: "orcamento",
  em_andamento: "em_andamento",
  finalizado: "finalizado",
  entregue: "entregue",
} as const;

export type ListDespesasParams = {
  dataInicio?: string;
  dataFim?: string;
};

export type GetResumoFinanceiroParams = {
  dataInicio?: string;
  dataFim?: string;
};

export type GetFluxoCaixaParams = {
  dataInicio?: string;
  dataFim?: string;
};