# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Oficina Mecânica Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCliente"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCliente"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # VEICULOS
...
  /financeiro/despesas:
    get:
      operationId: listDespesas
      tags: [financeiro]
      summary: List expenses
      parameters:
        - name: dataInicio
          in: query
          required: false
          schema:
            type: string
        - name: dataFim
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Despesa"
    post:
      operationId: createDespesa
      tags: [financeiro]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDespesa"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
  /financeiro/despesas/{id}:
    delete:
      operationId: deleteDespesa
      tags: [financeiro]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /financeiro/resumo:
    get:
      operationId: getResumoFinanceiro
      tags: [financeiro]
      summary: Get financial summary
      parameters:
        - name: dataInicio
          in: query
          required: false
          schema:
            type: string
        - name: dataFim
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumoFinanceiro"
  /financeiro/fluxo-caixa:
    get:
      operationId: getFluxoCaixa
      tags: [financeiro]
      summary: Get daily cash flow
      parameters:
        - name: dataInicio
          in: query
          required: false
          schema:
            type: string
        - name: dataFim
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Cash flow data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FluxoCaixaDia"
...
components:
  schemas:
...
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        observacoes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - nome
        - telefone
        - createdAt
    CreateCliente:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        observacoes:
          type: string
      required:
        - nome
        - telefone
...
    Despesa:
      type: object
      properties:
        id:
          type: integer
        descricao:
          type: string
        valor:
          type: number
        categoria:
          type: string
        data:
          type: string
        createdAt:
          type: string
      required:
        - id
        - descricao
        - valor
        - categoria
        - data
        - createdAt
    CreateDespesa:
      type: object
      properties:
        descricao:
          type: string
        valor:
          type: number
        categoria:
          type: string
        data:
          type: string
      required:
        - descricao
        - valor
        - categoria
        - data
    ResumoFinanceiro:
      type: object
      properties:
        totalReceitas:
          type: number
        totalDespesas:
          type: number
        lucro:
          type: number
        ordensFinalizadas:
          type: integer
        ticketMedio:
          type: number
      required:
        - totalReceitas
        - totalDespesas
        - lucro
        - ordensFinalizadas
        - ticketMedio
    FluxoCaixaDia:
      type: object
      properties:
        data:
          type: string
        receitas:
          type: number
        despesas:
          type: number
        saldo:
          type: number
      required:
        - data
        - receitas
        - despesas
        - saldo
...