# lib/api-zod/src/generated/types/createCliente.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Oficina Mecânica Management System
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCliente {
  nome: string;
  telefone: string;
  whatsapp?: string;
  email?: string;
  observacoes?: string;
}