# lib/api-zod/src/generated/types/despesa.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Oficina Mecânica Management System
 * OpenAPI spec version: 0.1.0
 */

export interface Despesa {
  id: number;
  descricao: string;
  valor: number;
  categoria: string;
  data: string;
  createdAt: string;
}